+++ modules/debugger/builtin/cpp/Gateway.cpp
@@ -17,20 +17,14 @@
 // LICENCE_BLOCK_END
 //=============================================================================
 #include "NelsonGateway.hpp"
-#include "iskeywordBuiltin.hpp"
-#include "max_recursion_depthBuiltin.hpp"
-#include "parsefileBuiltin.hpp"
-#include "parsestringBuiltin.hpp"
+#include "dbstackBuiltin.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
-const std::wstring gatewayName = L"interpreter";
+const std::wstring gatewayName = L"debugger";
 //=============================================================================
 static const nlsGateway gateway[] = {
-    { "iskeyword", Nelson::InterpreterGateway::iskeywordBuiltin, 1, 1 },
-    { "parsefile", Nelson::InterpreterGateway::parsefileBuiltin, 1, 1 },
-    { "parsestring", Nelson::InterpreterGateway::parsestringBuiltin, 1, 1 },
-    { "max_recursion_depth", Nelson::InterpreterGateway::max_recursion_depthBuiltin, 1, 1 },
+    { "dbstack", Nelson::DebuggerGateway::dbstackBuiltin, -1, -1 },
 };
 //=============================================================================
 NLSGATEWAYFUNC(gateway)