+++ modules/debugger/builtin/include/nlsDebugger_builtin_exports.h
@@ -16,25 +16,15 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-#include <Windows.h>
-#include "nlsConfig.h"
+#pragma once
 //=============================================================================
-#pragma comment(lib, CAT_3_STRINGS("boost_system-", BOOST_TARGET, ".lib"))
-#pragma comment(lib, CAT_3_STRINGS("boost_filesystem-", BOOST_TARGET, ".lib"))
-//=============================================================================
-int WINAPI
-DllMain(HINSTANCE hInstance, DWORD reason, PVOID pvReserved)
-{
-    switch (reason) {
-    case DLL_PROCESS_ATTACH:
-        break;
-    case DLL_PROCESS_DETACH:
-        break;
-    case DLL_THREAD_ATTACH:
-        break;
-    case DLL_THREAD_DETACH:
-        break;
-    }
-    return 1;
-}
+#ifdef _MSC_VER
+#ifdef NLSDEBUGGER_BUILTIN_EXPORTS
+#define NLSDEBUGGER_BUILTIN_IMPEXP __declspec(dllexport)
+#else
+#define NLSDEBUGGER_BUILTIN_IMPEXP __declspec(dllimport)
+#endif
+#else
+#define NLSDEBUGGER_BUILTIN_IMPEXP __attribute__((visibility("default")))
+#endif
 //=============================================================================