+++ CHANGELOG.md
@@ -49,6 +49,9 @@ Features:
 
   * [#159](http://github.com/Nelson-numerical-software/nelson/issues/159): addpath must return an warning and not an error for an non existing path.
 
+  * [#156](http://github.com/Nelson-numerical-software/nelson/issues/156): Moves dbstack in 'debugger' module.
+
+
 
 ## 0.2.9 (2018-09-23)